refactor: guard tauri log plugin
